Finished up the state engine, added a title screen.
